Move CONF_ROWS to components/const/
