+++ esphome/components/const/__init__.py
@@ -12,3 +12,4 @@ CONF_ON_RECEIVE = "on_receive"
 CONF_ON_STATE_CHANGE = "on_state_change"
 CONF_REQUEST_HEADERS = "request_headers"
 CONF_USE_PSRAM = "use_psram"
+CONF_ROWS = "rows"
