+++ esphome/components/matrix_keypad/__init__.py
@@ -1,8 +1,9 @@
 from esphome import automation, pins
 import esphome.codegen as cg
 from esphome.components import key_provider
+from esphome.components.const import CONF_ROWS
 import esphome.config_validation as cv
-from esphome.const import CONF_ID, CONF_ON_KEY, CONF_PIN, CONF_ROWS, CONF_TRIGGER_ID
+from esphome.const import CONF_ID, CONF_ON_KEY, CONF_PIN, CONF_TRIGGER_ID
 
 CODEOWNERS = ["@ssieb"]
